+++ src/features/notification/api/notification.API.ts
@@ -1,6 +1,6 @@
-import axiosInstance from '@/shared/api/axios-instance'
+import { axiosInstance } from '@/shared/api'
 
-import type { NotificationResponse } from '@/entities/notification/models/notification.types'
+import type { NotificationResponse } from '@/entities/notification/models'
 
 export async function getNotification(page = 1, size = 10): Promise<NotificationResponse> {
   const { data } = await axiosInstance.get('/notification', {
